Fixed type for maxIntXXX

+++ cpp/daal/src/sycl/select_indexed.cpp
@@ -29,7 +29,7 @@
 using namespace daal::data_management;
 using namespace daal::services::internal;
 
-constexpr size_t maxInt32AsUint32T = static_cast<uint32_t>(daal::services::internal::MaxVal<int32_t>::get());
+constexpr uint32_t maxInt32AsUint32T = static_cast<uint32_t>(daal::services::internal::MaxVal<int32_t>::get());
 
 namespace daal
 {
